@@ -374,7 +374,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
             var extendedDataCount = data.eventRecord->ExtendedDataCount;
             if (extendedDataCount != 0)
             {
-                bookKeepingEvent |= ProcessExtendedData(data, extendedDataCount, countForEvent);
+                bookKeepingEvent |= ProcessExtendedData(data, extendedDataCount, countForEvent, isLiveSession: true);
             }
 
             // This must occur after the call to ProcessExtendedData to ensure that if there is a stack for this event,
@@ -858,7 +858,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                 var extendedDataCount = data.eventRecord->ExtendedDataCount;
                 if (extendedDataCount != 0)
                 {
-                    bookKeepingEvent |= ProcessExtendedData(data, extendedDataCount, countForEvent);
+                    bookKeepingEvent |= ProcessExtendedData(data, extendedDataCount, countForEvent, isLiveSession: true);
                 }
 
                 // This must occur after the call to ProcessExtendedData to ensure that if there is a stack for this event,
@@ -2269,7 +2269,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                 var extendedDataCount = data.eventRecord->ExtendedDataCount;
                 if (extendedDataCount != 0)
                 {
-                    bookKeepingEvent |= ProcessExtendedData(data, extendedDataCount, countForEvent);
+                    bookKeepingEvent |= ProcessExtendedData(data, extendedDataCount, countForEvent, isLiveSession: false);
                 }
 
                 if (bookKeepingEvent)
@@ -3245,7 +3245,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
         /// Process any extended data (like Win7 style stack traces) associated with 'data'
         /// returns true if the event should be considered a bookkeeping event.
         /// </summary>
-        internal unsafe bool ProcessExtendedData(TraceEvent data, ushort extendedDataCount, TraceEventCounts countForEvent)
+        internal unsafe bool ProcessExtendedData(TraceEvent data, ushort extendedDataCount, TraceEventCounts countForEvent, bool isLiveSession)
         {
             var isBookkeepingEvent = false;
             var extendedData = data.eventRecord->ExtendedData;
@@ -3334,72 +3334,79 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                         }
                     }
                 }
-                else if (extendedData[i].ExtType == TraceEventNativeMethods.EVENT_HEADER_EXT_TYPE_RELATED_ACTIVITYID)
+                // In live sessions, preserve the original ExtendedData for readers to access directly.
+                // Only persist RelatedActivityID in file conversion mode.
+                else if (!isLiveSession && extendedData[i].ExtType == TraceEventNativeMethods.EVENT_HEADER_EXT_TYPE_RELATED_ACTIVITYID)
                 {
                     relatedActivityIDPtr = (Guid*)(extendedData[i].DataPtr);
                 }
-                else if (extendedData[i].ExtType == TraceEventNativeMethods.EVENT_HEADER_EXT_TYPE_CONTAINER_ID)
+                else if (!isLiveSession && extendedData[i].ExtType == TraceEventNativeMethods.EVENT_HEADER_EXT_TYPE_CONTAINER_ID)
                 {
                     containerID = Marshal.PtrToStringAnsi((IntPtr)extendedData[i].DataPtr, (int)extendedData[i].DataSize);
                 }
             }
 
-            if (relatedActivityIDPtr != null)
+            // Only persist RelatedActivityID and modify ExtendedData during file conversion.
+            // In live sessions, preserve the original ExtendedData pointers for readers to access.
+            if (!isLiveSession)
             {
-                if (relatedActivityIDs.Count == 0)
+                if (relatedActivityIDPtr != null)
                 {
-                    // Insert a synthetic value since 0 represents "no related activity ID".
-                    relatedActivityIDs.Add(Guid.Empty);
-                }
+                    if (relatedActivityIDs.Count == 0)
+                    {
+                        // Insert a synthetic value since 0 represents "no related activity ID".
+                        relatedActivityIDs.Add(Guid.Empty);
+                    }
 
-                // TODO This is a bit of a hack.   We wack this field in place.
-                // We encode this as index into the relatedActivityID GrowableArray.
-                if (IntPtr.Size == 8)
-                {
-                    data.eventRecord->ExtendedData = (TraceEventNativeMethods.EVENT_HEADER_EXTENDED_DATA_ITEM*)(relatedActivityIDs.Count << 4);
+                    // TODO This is a bit of a hack.   We wack this field in place.
+                    // We encode this as index into the relatedActivityID GrowableArray.
+                    if (IntPtr.Size == 8)
+                    {
+                        data.eventRecord->ExtendedData = (TraceEventNativeMethods.EVENT_HEADER_EXTENDED_DATA_ITEM*)(relatedActivityIDs.Count << 4);
+                    }
+                    else
+                    {
+                        data.eventRecord->ExtendedData = (TraceEventNativeMethods.EVENT_HEADER_EXTENDED_DATA_ITEM*)relatedActivityIDs.Count;
+                    }
+                    relatedActivityIDs.Add(*relatedActivityIDPtr);
                 }
                 else
                 {
-                    data.eventRecord->ExtendedData = (TraceEventNativeMethods.EVENT_HEADER_EXTENDED_DATA_ITEM*)relatedActivityIDs.Count;
-                }
-                relatedActivityIDs.Add(*relatedActivityIDPtr);
-            }
-            else
-            {
-                data.eventRecord->ExtendedData = null;
-            }
-
-            if (containerID != null)
-            {
-                // TODO This is a bit of a hack.   We wack this field in place.
-                // We encode this as index into the containerIDs GrowableArray.
-                if (containerIDs.Count == 0)
-                {
-                    // Insert a synthetic value since 0 represents "no container ID".
-                    containerIDs.Add(null);
+                    data.eventRecord->ExtendedData = null;
                 }
 
-                // Look for the container ID.
-                bool found = false;
-                for (int i = 0; i < containerIDs.Count; i++)
+                if (containerID != null)
                 {
-                    if (containerIDs[i] == containerID)
+                    // TODO This is a bit of a hack.   We wack this field in place.
+                    // We encode this as index into the containerIDs GrowableArray.
+                    if (containerIDs.Count == 0)
                     {
-                        data.eventRecord->ExtendedDataCount = (ushort)i;
-                        found = true;
-                        break;
+                        // Insert a synthetic value since 0 represents "no container ID".
+                        containerIDs.Add(null);
+                    }
+
+                    // Look for the container ID.
+                    bool found = false;
+                    for (int i = 0; i < containerIDs.Count; i++)
+                    {
+                        if (containerIDs[i] == containerID)
+                        {
+                            data.eventRecord->ExtendedDataCount = (ushort)i;
+                            found = true;
+                            break;
+                        }
+                    }
+
+                    if (!found)
+                    {
+                        data.eventRecord->ExtendedDataCount = (ushort)containerIDs.Count;
+                        containerIDs.Add(containerID);
                     }
                 }
-
-                if (!found)
+                else
                 {
-                    data.eventRecord->ExtendedDataCount = (ushort)containerIDs.Count;
-                    containerIDs.Add(containerID);
+                    data.eventRecord->ExtendedDataCount = 0;
                 }
-            }
-            else
-            {
-                data.eventRecord->ExtendedDataCount = 0;
             }
 
             return isBookkeepingEvent;
@@ -3795,6 +3802,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
             return true;
         }
 #endif
+
         void IFastSerializable.ToStream(Serializer serializer)
         {
             // Write out the events themselves, Before we do this we write a reference past the end of the
@@ -4725,6 +4733,19 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
 
             sb.AppendLine("</TraceEventStats>");
             return sb.ToString();
+        }
+
+        internal bool Contains(Guid providerId, string eventNameSearchVal)
+        {
+            foreach (var counts in m_counts.Values)
+            {
+                if (counts.ProviderGuid == providerId &&
+                    counts.EventName.Contains(eventNameSearchVal))
+                {
+                    return true;
+                }
+            }
+            return false;
         }
 
         #region private
@@ -7106,31 +7127,47 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
 
         internal TraceModuleFile UniversalMapping(ProcessMappingTraceData data, ProcessMappingMetadataTraceData metadata)
         {
+            return UniversalMapping(data.FileName, data.StartAddress, data.EndAddress, data.TimeStampQPC, metadata);
+        }
+
+        internal TraceModuleFile UniversalMapping(string fileName, Address startAddress, Address endAddress, long timeStampQPC, ProcessMappingMetadataTraceData metadata)
+        {
             int index;
 
             // A loaded and managed modules depend on a module file, so get or create one.
+            // The key is the file name.  For jitted code on Linux, this will be a memfd with a static name, which is OK
+            // because this path will use the StartAddress to ensure that we get the right one.
             // TODO: We'll need to store FileOffset as well to handle elf images.
-            TraceModuleFile moduleFile = process.Log.ModuleFiles.GetOrCreateModuleFile(data.FileName, data.StartAddress);
-            moduleFile.imageSize = (long)(data.EndAddress - data.StartAddress);
+            TraceModuleFile moduleFile = process.Log.ModuleFiles.GetOrCreateModuleFile(fileName, startAddress);
+            long newImageSize = (long)(endAddress - startAddress);
+            
+            // New mappings will have an imageSize of 0 and will get set.
+            // Existing mappings that have the same StartAddress but increase in length will get updated here.
+            if (moduleFile.imageSize < newImageSize)
+            {
+                moduleFile.imageSize = newImageSize;
+            }
 
-            // Get or create the loaded module.
-            TraceLoadedModule loadedModule = FindModuleAndIndexContainingAddress(data.StartAddress, data.TimeStampQPC, out index);
-            if (loadedModule == null)
-            {   
+            // The loaded module is looked up by StartAddress and time to ensure that we don't use a module that hasn't been loaded yet.
+            // If the StartAddress or size don't match, then create a new one.  This handles overlapping cases.
+            TraceLoadedModule loadedModule = FindModuleAndIndexContainingAddress(startAddress, timeStampQPC, out index);
+            if (loadedModule == null || loadedModule.ImageBase != startAddress || loadedModule.ModuleFile.imageSize != newImageSize)
+            {
                 // The module file is what is used when looking up the module for an arbitrary address, so it must save both the start address and image size.
-                loadedModule = new TraceLoadedModule(process, moduleFile, data.StartAddress);
-                
-                // All mappings are enumerated at the beginning of the trace.
-                loadedModule.loadTimeQPC = process.Log.sessionStartTimeQPC;
-                
+                loadedModule = new TraceLoadedModule(process, moduleFile, startAddress);
+
+                // Set the timestamp from the mapping data
+                loadedModule.loadTimeQPC = timeStampQPC;
+
                 InsertAndSetOverlap(index + 1, loadedModule);
             }
 
             // Get or create a managed module.  This module is the container for dynamic symbols.
-            TraceManagedModule managedModule = FindManagedModuleAndIndex((long)data.StartAddress, data.TimeStampQPC, out index);
+            TraceManagedModule managedModule = FindManagedModuleAndIndex((long)startAddress, timeStampQPC, out index);
             if (managedModule == null)
             {
-                managedModule = new TraceManagedModule(process, moduleFile, (long)data.StartAddress);
+                managedModule = new TraceManagedModule(process, moduleFile, (long)startAddress);
+                managedModule.loadTimeQPC = timeStampQPC;
                 modules.Insert(index + 1, managedModule);
             }
 
@@ -8524,16 +8561,39 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                     if (module == null)
                     {
                         int index;
+                        string moduleName = "UNKNOWN";
+                        string methodName = data.Name;
                         TraceLoadedModule loadedModule = process.LoadedModules.FindModuleAndIndexContainingAddress(data.StartAddress, data.TimeStampQPC, out index);
+
+                        // Try to parse the symbol as a universal symbol.
+                        var parsed = NettraceUniversalConverter.ParseDotnetJittedSymbolName(data.Name);
+                        if (parsed.HasValue)
+                        {
+                            moduleName = parsed.Value.moduleName;
+                            methodName = parsed.Value.methodSignature;
+                        }
+
+                        // We don't create a blanket jitted code module, so create one here.
+                        // Non-jitted symbols will already have a module, so loadedModule will not be null.
+                        if (loadedModule == null)
+                        {
+                            TraceModuleFile moduleFile = process.LoadedModules.UniversalMapping(moduleName, data.StartAddress, data.EndAddress, data.TimeStampQPC, null);
+                            loadedModule = process.LoadedModules.FindModuleAndIndexContainingAddress(data.StartAddress, data.TimeStampQPC, out index);
+                        }
+
                         module = process.LoadedModules.GetOrCreateManagedModule(loadedModule.ModuleID, data.TimeStampQPC);
                         moduleFileIndex = module.ModuleFile.ModuleFileIndex;
-                        methodIndex = methods.NewMethod(data.Name, moduleFileIndex, (int)data.Id);
+                            
+                        methodIndex = methods.NewMethod(methodName, moduleFileIndex, (int)data.Id);
                         
                         // When universal traces support re-use of address space, we'll need to support it here.
                     }
 
-                    // Set the info
-                    info.SetMethodIndex(this, methodIndex);
+                    // Set the info (only if we found a module)
+                    if (module != null)
+                    {
+                        info.SetMethodIndex(this, methodIndex);
+                    }
                 }
             });
         }
@@ -9079,7 +9139,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
                     else
                     {
                         symReader.m_log.WriteLine("ERROR: The R2R perfmap does not match the loaded module.  Actual Signature = " + symbolModule.Signature + " Requested Signature = " + moduleFile.R2RPerfMapSignature);
-                        return null;
+                        throw new Exception("ERROR: The R2R perfmap does not match the loaded module.");
                     }
                 }
             }
